Change int16_t to int to improve speed

+++ brute_force.hpp
@@ -6,6 +6,6 @@
 // them in the array `output`.
 // This brute force solution is only used to check correctness.
 int brute_force_search(
-    const char *text, int text_length, const char *pattern, int16_t pattern_length,
-    int *output, int max_output_cnt, int16_t *fail
+    const char *text, int text_length, const char *pattern, int pattern_length,
+    int *output, int max_output_cnt, int *fail
 );
